requirements.txt
fastapi==0.78.0
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dateutil==2.8.2
SQLAlchemy==1.4.39